requirements.txt
# Django
# https://www.djangoproject.com/
Django==2.0.6

# Postgres conector
# http://initd.org/psycopg/
psycopg2==2.7.4

# Django Environ
# http://django-environ.readthedocs.io/en/latest/
django-environ==0.4.4

# Django CORS Headers
# https://github.com/ottoyiu/django-cors-headers
django-cors-headers==2.2.0

# Django REST Framework
# http://www.django-rest-framework.org/
djangorestframework==3.8.2

# Django Filters
# http://django-filter.readthedocs.io/en/latest/
django-filter==1.1.0

# Django Extensions
# https://django-extensions.readthedocs.io/en/latest/
django-extensions==2.0.7
